Updates the ArgumentDefinition to use AttributeType

+++ core/nmf-composites/src/main/java/esa/mo/nmf/annotations/ActionParameter.java
@@ -38,13 +38,13 @@ public @interface ActionParameter {
 
     String description() default "";
 
-    byte rawType() default 0;
+    int rawType() default 0;
 
     String rawUnit() default "";
 
     String conditionalConversionFieldName() default "";
 
-    byte convertedType() default -1;
+    int convertedType() default -1;
 
     String convertedUnit() default "";
 }
